Changed naming and formatted files
