+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ImageUploader from './component/UploadImage/UploadSection';
+import ImageUploader from './component/ImageUploader/ImageUploader';
 const InitialComponent: React.FC = () => {
     return (
         <ImageUploader/>
